backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
bcrypt==4.1.3
shapely==2.0.6
geojson==3.1.0
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
black==24.10.0
ruff==0.6.9
isort==5.13.2
mypy==1.10.0
